sediment: cleanup
