+++ frontend/src/components/ClarificationModal.tsx
@@ -84,7 +84,7 @@ export function ClarificationModal({ question, options, onSelect, onDismiss }: P
                 style={{
                   fontSize: "0.625rem",
                   color: "var(--accent)",
-                  fontFamily: "'Geist Mono', monospace",
+                  fontFamily: "var(--font-mono), monospace",
                   letterSpacing: "0.1em",
                   textTransform: "uppercase",
                   marginBottom: "0.5rem",
@@ -97,7 +97,7 @@ export function ClarificationModal({ question, options, onSelect, onDismiss }: P
                 style={{
                   fontSize: "1rem",
                   color: "var(--text-primary)",
-                  fontFamily: "'Inter', sans-serif",
+                  fontFamily: "var(--font-sans), sans-serif",
                   fontWeight: 500,
                   lineHeight: 1.4,
                 }}
@@ -143,7 +143,7 @@ export function ClarificationModal({ question, options, onSelect, onDismiss }: P
                     background: "var(--bg-secondary)",
                     color: "var(--text-primary)",
                     fontSize: "0.875rem",
-                    fontFamily: "'Inter', sans-serif",
+                    fontFamily: "var(--font-sans), sans-serif",
                     fontWeight: 500,
                     cursor: "pointer",
                     transition: "border-color 0.12s, background 0.12s, color 0.12s",
@@ -174,7 +174,7 @@ export function ClarificationModal({ question, options, onSelect, onDismiss }: P
           {/* Divider */}
           <div style={{ display: "flex", alignItems: "center", gap: "0.625rem" }}>
             <div style={{ flex: 1, height: "0.0625rem", background: "var(--border)" }} />
-            <span style={{ fontSize: "0.6875rem", color: "var(--text-tertiary)", fontFamily: "'Geist Mono', monospace", letterSpacing: "0.04em" }}>
+            <span style={{ fontSize: "0.6875rem", color: "var(--text-tertiary)", fontFamily: "var(--font-mono), monospace", letterSpacing: "0.04em" }}>
               or type your own
             </span>
             <div style={{ flex: 1, height: "0.0625rem", background: "var(--border)" }} />
@@ -197,7 +197,7 @@ export function ClarificationModal({ question, options, onSelect, onDismiss }: P
                 background: "var(--bg-primary)",
                 color: "var(--text-primary)",
                 fontSize: "0.875rem",
-                fontFamily: "'Inter', sans-serif",
+                fontFamily: "var(--font-sans), sans-serif",
                 outline: "none",
                 transition: "border-color 0.12s",
               }}
@@ -215,7 +215,7 @@ export function ClarificationModal({ question, options, onSelect, onDismiss }: P
                 background: "var(--accent-soft)",
                 color: "var(--accent)",
                 fontSize: "0.8125rem",
-                fontFamily: "'Inter', sans-serif",
+                fontFamily: "var(--font-sans), sans-serif",
                 fontWeight: 600,
                 cursor: customValue.trim() ? "pointer" : "default",
                 opacity: customValue.trim() ? 1 : 0.45,
